+++ packages/contracts/src/index.ts
@@ -14,3 +14,4 @@ export {
 export * from './rbac';
 export * from './cit';
 export * from './types';
+export * from './constants';